requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-dotenv==1.0.1
google-generativeai==0.8.0
httpx==0.27.0
pydantic==2.6.1
PyPDF2==3.0.1
scikit-learn==1.5.0
numpy==1.26.4
beautifulsoup4==4.12.3
aiofiles==23.2.1
python-multipart==0.0.9
starlette==0.36.3
filelock==3.13.1
fsspec==2024.3.1
networkx==3.2.1
sympy==1.12
packaging==23.2
regex==2023.12.25
safetensors==0.4.3
huggingface-hub==0.23.4
pdfminer.six==20231228
deep-translator==1.11.4
